Moved the component library server logic  into its own class.
